# refactor(query): Improve parse json performance