I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary

This PR optimizes `parse_json` performance by replacing the `BTreeMap` used for parsing JSON objects with a `Vec`.  The new implementation avoids the overhead of `BTreeMap` insertions, resulting in significant speedups.

Parsing large JSON objects shows substantial improvement:

*   **10M rows:** 10.617s -> 7.242s (~31% faster)
*   **100M rows:** 185.819s -> 73.034s (~60% faster)

- fixes: #[Link the issue here]


## Tests

- [x] Unit Test
- [ ] Logic Test
- [ ] Benchmark Test
- [ ] No Test - _Explain why_

## Type of change

- [ ] Bug Fix (non-breaking change which fixes an issue)
- [ ] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [ ] Refactoring
- [x] Performance Improvement
- [ ] Other (please describe):

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/databendlabs/databend/18451)
<!-- Reviewable:end -->